sonarScanner working, when publishing arrays is discovered it should look boss. dwncam2 is what I was using for testing, currently has some sample color segmentation code on

+++ Soft_Nodes/sonarScanner/src/sonarScanner.h
@@ -2,7 +2,7 @@
 float bins;
 float bearing;
 
-void pixelPlace( unsigned int theta, unsigned int distance, unsigned opaqueVal );
+void pixelPlace( float theta, unsigned int distance, unsigned opaqueVal );
 int genRand( int n );
 void printascii( void );
 void drawScene(unsigned int x, unsigned int y);
